refactor: split client.ts into 7 domain files (6,062 → 207 line barrel)
Complete Phase 2: MiladyClient method splitting via TypeScript declaration
merging + prototype augmentation. Core infrastructure (constructor, fetch,
rawRequest, WS lifecycle) stays in client-base.ts as protected methods.
Five domain files augment the prototype with their method groups.

client.ts:       6,062 → 207 lines (thin barrel + singleton)
client-base.ts:  754 lines (core class infrastructure)
client-types.ts: 1,848 lines (all types/interfaces)
client-agent.ts: 1,717 lines (agent, auth, config, plugins, training, PTY)
client-chat.ts:  1,000 lines (chat, knowledge, memory, workbench, DB)
client-cloud.ts: 674 lines (cloud billing, compat agents, sandbox)
client-skills.ts: 715 lines (skills, catalog, marketplace, apps)
client-wallet.ts: 419 lines (wallet, trading, steward, registry)

Zero consumer changes — barrel re-exports everything.
