+++ packages/app-core/src/api/client-base.ts
@@ -1,0 +1,754 @@
+/**
+ * MiladyClient class — core infrastructure only.
+ *
+ * Separated from client.ts so domain augmentation files can import the class
+ * without circular dependency issues.
+ */
+
+import { stripAssistantStageDirections } from "../utils/assistant-text";
+import { getElizaApiBase, getElizaApiToken } from "../utils/eliza-globals";
+import { mergeStreamingText } from "../utils/streaming-text";
+import { getBootConfig, setBootConfig } from "../config/boot-config";
+import type {
+  ChatTokenUsage,
+  ConnectionStateInfo,
+  ConversationChannelType,
+  ConversationMessage,
+  ConversationMode,
+  ImageAttachment,
+  WebSocketConnectionState,
+  WsEventHandler,
+} from "./client-types";
+import { ApiError } from "./client-types";
+
+// ---------------------------------------------------------------------------
+// Constants
+// ---------------------------------------------------------------------------
+
+const GENERIC_NO_RESPONSE_TEXT =
+  "Sorry, I couldn't generate a response right now. Please try again.";
+const DEFAULT_FETCH_TIMEOUT_MS = 10_000;
+const SESSION_STORAGE_API_BASE_KEY = "milady_api_base";
+
+// ---------------------------------------------------------------------------
+// Client
+// ---------------------------------------------------------------------------
+
+export class MiladyClient {
+  private _baseUrl: string;
+  private _explicitBase: boolean;
+  private _token: string | null;
+  private readonly clientId: string;
+  private ws: WebSocket | null = null;
+  private wsHandlers = new Map<string, Set<WsEventHandler>>();
+  private wsSendQueue: string[] = [];
+  private readonly wsSendQueueLimit = 32;
+  private reconnectTimer: ReturnType<typeof setTimeout> | null = null;
+  private backoffMs = 500;
+  private wsHasConnectedOnce = false;
+
+  // Connection state tracking for backend crash handling
+  private connectionState: WebSocketConnectionState = "disconnected";
+  private reconnectAttempt = 0;
+  private disconnectedAt: number | null = null;
+  private connectionStateListeners = new Set<
+    (state: ConnectionStateInfo) => void
+  >();
+  private readonly maxReconnectAttempts = 15;
+
+  // UI language propagation — set by AppContext so the backend can
+  // localise responses when needed.
+  private _uiLanguage: string | null = null;
+
+  /** Store the current UI language so it can be sent as a header on every request. */
+  setUiLanguage(lang: string): void {
+    this._uiLanguage = lang || null;
+  }
+
+  private static generateClientId(): string {
+    const random =
+      typeof globalThis.crypto?.randomUUID === "function"
+        ? globalThis.crypto.randomUUID()
+        : `${Date.now().toString(36)}${Math.random().toString(36).slice(2)}`;
+    return `ui-${random.replace(/[^a-zA-Z0-9._-]/g, "")}`;
+  }
+
+  constructor(baseUrl?: string, token?: string) {
+    this.clientId = MiladyClient.generateClientId();
+    this._token = token?.trim() || null;
+
+    const injectedBase = getElizaApiBase();
+    const storedBase =
+      typeof window !== "undefined"
+        ? window.sessionStorage.getItem(SESSION_STORAGE_API_BASE_KEY)
+        : null;
+    const bootBase = getBootConfig().apiBase;
+
+    this._explicitBase = baseUrl != null || Boolean(storedBase?.trim());
+
+    // Priority: explicit arg > desktop injection > boot config > session storage > same origin.
+    // Injected global (window.__MILADY_API_BASE__) must beat stale sessionStorage
+    // from a prior cloud/remote session so the desktop always connects to the
+    // correct local agent.
+    this._baseUrl =
+      baseUrl ?? injectedBase ?? bootBase ?? storedBase ?? "";
+  }
+
+  /**
+   * Resolve the API base URL lazily.
+   * In the desktop shell the main process injects the API base after the
+   * page loads (once the agent runtime starts). Re-checking the boot config
+   * on every call ensures we pick up the injected value even if it wasn't
+   * set at construction.
+   */
+  protected get baseUrl(): string {
+    if (!this._explicitBase) {
+      const bootBase = getBootConfig().apiBase ?? getElizaApiBase();
+      if (bootBase && bootBase !== this._baseUrl) {
+        this._baseUrl = bootBase;
+      }
+    }
+    return this._baseUrl;
+  }
+
+  protected get apiToken(): string | null {
+    if (this._token) return this._token;
+    const bootToken = getBootConfig().apiToken;
+    if (typeof bootToken === "string" && bootToken.trim())
+      return bootToken.trim();
+    const injectedToken = getElizaApiToken();
+    if (injectedToken) return injectedToken;
+    return null;
+  }
+
+  hasToken(): boolean {
+    return Boolean(this.apiToken);
+  }
+
+  /**
+   * Bearer token sent on Milady REST requests (compat API). Used when the
+   * Electrobun main process relays HTTP so it can match the renderer-injected
+   * token in external-desktop / Vite-proxy setups.
+   */
+  getRestAuthToken(): string | null {
+    return this.apiToken;
+  }
+
+  setToken(token: string | null): void {
+    this._token = token?.trim() || null;
+    // Update boot config so other consumers see the new token.
+    const config = getBootConfig();
+    setBootConfig({ ...config, apiToken: this._token ?? undefined });
+  }
+
+  getBaseUrl(): string {
+    return this.baseUrl;
+  }
+
+  setBaseUrl(baseUrl: string | null): void {
+    const normalized = baseUrl?.trim().replace(/\/+$/, "") || "";
+    this._explicitBase = normalized.length > 0;
+    this._baseUrl = normalized;
+    this.disconnectWs();
+    // Update boot config so other consumers (resolveApiUrl, etc.) see the new base.
+    const config = getBootConfig();
+    setBootConfig({ ...config, apiBase: normalized || undefined });
+    if (typeof window !== "undefined") {
+      if (normalized) {
+        window.sessionStorage.setItem(SESSION_STORAGE_API_BASE_KEY, normalized);
+      } else {
+        window.sessionStorage.removeItem(SESSION_STORAGE_API_BASE_KEY);
+      }
+    }
+  }
+
+  /** True when we have a usable HTTP(S) API endpoint. */
+  get apiAvailable(): boolean {
+    if (this.baseUrl) return true;
+    if (typeof window !== "undefined") {
+      const proto = window.location.protocol;
+      return proto === "http:" || proto === "https:";
+    }
+    return false;
+  }
+
+  // --- REST API ---
+
+  protected async rawRequest(
+    path: string,
+    init?: RequestInit,
+    options?: { allowNonOk?: boolean },
+  ): Promise<Response> {
+    if (!this.apiAvailable) {
+      throw new ApiError({
+        kind: "network",
+        path,
+        message: "API not available (no HTTP origin)",
+      });
+    }
+    const makeRequest = async (token: string | null): Promise<Response> => {
+      let timeoutId: ReturnType<typeof setTimeout> | undefined;
+      let abortListener: (() => void) | undefined;
+      const requestInit: RequestInit = {
+        ...init,
+        headers: {
+          "X-Milady-Client-Id": this.clientId,
+          ...(token ? { Authorization: `Bearer ${token}` } : {}),
+          ...(this._uiLanguage
+            ? { "X-Milady-UI-Language": this._uiLanguage }
+            : {}),
+          ...init?.headers,
+        },
+      };
+
+      const fetchPromise = fetch(`${this.baseUrl}${path}`, requestInit);
+      const pending: Promise<Response>[] = [fetchPromise];
+
+      pending.push(
+        new Promise<Response>((_, reject) => {
+          timeoutId = setTimeout(() => {
+            reject(
+              new ApiError({
+                kind: "timeout",
+                path,
+                message: `Request timed out after ${DEFAULT_FETCH_TIMEOUT_MS}ms`,
+              }),
+            );
+          }, DEFAULT_FETCH_TIMEOUT_MS);
+        }),
+      );
+
+      if (init?.signal) {
+        if (init.signal.aborted) {
+          throw new ApiError({
+            kind: "network",
+            path,
+            message: "Request aborted",
+          });
+        }
+
+        pending.push(
+          new Promise<Response>((_, reject) => {
+            abortListener = () => {
+              reject(
+                new ApiError({
+                  kind: "network",
+                  path,
+                  message: "Request aborted",
+                }),
+              );
+            };
+            init.signal?.addEventListener("abort", abortListener, {
+              once: true,
+            });
+          }),
+        );
+      }
+
+      try {
+        return await Promise.race(pending);
+      } catch (err) {
+        if (err instanceof ApiError) {
+          throw err;
+        }
+        throw new ApiError({
+          kind: "network",
+          path,
+          message:
+            err instanceof Error && err.message
+              ? err.message
+              : "Network request failed",
+          cause: err,
+        });
+      } finally {
+        if (timeoutId !== undefined) {
+          clearTimeout(timeoutId);
+        }
+        if (init?.signal && abortListener) {
+          init.signal.removeEventListener("abort", abortListener);
+        }
+      }
+    };
+
+    const token = this.apiToken;
+    let res = await makeRequest(token);
+    if (res.status === 401 && !token) {
+      const retryToken = this.apiToken;
+      if (retryToken) {
+        res = await makeRequest(retryToken);
+      }
+    }
+    if (!res.ok && !options?.allowNonOk) {
+      const body = (await res
+        .json()
+        .catch(() => ({ error: res.statusText }))) as Record<
+        string,
+        string
+      > | null;
+      throw new ApiError({
+        kind: "http",
+        path,
+        status: res.status,
+        message: body?.error ?? `HTTP ${res.status}`,
+      });
+    }
+    return res;
+  }
+
+  protected async fetch<T>(path: string, init?: RequestInit): Promise<T> {
+    const res = await this.rawRequest(path, {
+      ...init,
+      headers: {
+        "Content-Type": "application/json",
+        ...init?.headers,
+      },
+    });
+    return res.json() as Promise<T>;
+  }
+
+  // --- WebSocket ---
+
+  connectWs(): void {
+    if (this.ws?.readyState === WebSocket.OPEN) return;
+
+    let host: string;
+    let wsProtocol: "ws:" | "wss:";
+    if (this.baseUrl) {
+      const parsed = new URL(this.baseUrl);
+      host = parsed.host;
+      wsProtocol = parsed.protocol === "https:" ? "wss:" : "ws:";
+    } else {
+      // In non-HTTP environments (electrobun://, file://, etc.)
+      // window.location.host may be empty or a non-routable placeholder like "-".
+      const loc = window.location;
+      if (loc.protocol !== "http:" && loc.protocol !== "https:") return;
+      host = loc.host;
+      wsProtocol = loc.protocol === "https:" ? "wss:" : "ws:";
+    }
+
+    if (!host) return;
+
+    let url = `${wsProtocol}//${host}/ws`;
+    const params = new URLSearchParams({ clientId: this.clientId });
+    url += `?${params.toString()}`;
+
+    this.ws = new WebSocket(url);
+
+    this.ws.onopen = () => {
+      const token = this.apiToken;
+      if (token && this.ws?.readyState === WebSocket.OPEN) {
+        this.ws.send(JSON.stringify({ type: "auth", token }));
+      }
+      this.backoffMs = 500;
+      // Reset connection state on successful connection
+      this.reconnectAttempt = 0;
+      this.disconnectedAt = null;
+      this.connectionState = "connected";
+      this.emitConnectionStateChange();
+
+      // Notify listeners when the WS reconnects (not on the first connect)
+      // so they can re-hydrate state that may have been lost during the gap.
+      if (this.wsHasConnectedOnce) {
+        const handlers = this.wsHandlers.get("ws-reconnected");
+        if (handlers) {
+          for (const handler of handlers) {
+            handler({ type: "ws-reconnected" });
+          }
+        }
+      }
+      this.wsHasConnectedOnce = true;
+      if (
+        this.wsSendQueue.length > 0 &&
+        this.ws?.readyState === WebSocket.OPEN
+      ) {
+        const pending = this.wsSendQueue;
+        this.wsSendQueue = [];
+        for (let i = 0; i < pending.length; i++) {
+          if (this.ws?.readyState !== WebSocket.OPEN) {
+            this.wsSendQueue = pending.slice(i).concat(this.wsSendQueue);
+            break;
+          }
+          try {
+            this.ws.send(pending[i]);
+          } catch {
+            this.wsSendQueue = pending.slice(i).concat(this.wsSendQueue);
+            break;
+          }
+        }
+      }
+    };
+
+    this.ws.onmessage = (event) => {
+      try {
+        const data = JSON.parse(event.data as string) as Record<
+          string,
+          unknown
+        >;
+        const type = data.type as string;
+        const handlers = this.wsHandlers.get(type);
+        if (handlers) {
+          for (const handler of handlers) {
+            handler(data);
+          }
+        }
+        // Also fire "all" handlers
+        const allHandlers = this.wsHandlers.get("*");
+        if (allHandlers) {
+          for (const handler of allHandlers) {
+            handler(data);
+          }
+        }
+      } catch {
+        // ignore parse errors
+      }
+    };
+
+    this.ws.onclose = () => {
+      this.ws = null;
+      // Track disconnection time if not already set
+      if (this.disconnectedAt === null) {
+        this.disconnectedAt = Date.now();
+      }
+      this.reconnectAttempt++;
+      // Update state based on attempt count
+      if (this.reconnectAttempt >= this.maxReconnectAttempts) {
+        this.connectionState = "failed";
+      } else {
+        this.connectionState = "reconnecting";
+      }
+      this.emitConnectionStateChange();
+      this.scheduleReconnect();
+    };
+
+    this.ws.onerror = () => {
+      // close handler will fire
+    };
+  }
+
+  private scheduleReconnect(): void {
+    if (this.reconnectTimer) return;
+    // After the short backoff window is exhausted, keep probing at a
+    // low frequency so the UI can recover without a full page refresh.
+    if (this.reconnectAttempt >= this.maxReconnectAttempts) {
+      this.reconnectTimer = setTimeout(() => {
+        this.reconnectTimer = null;
+        this.connectWs();
+      }, 30_000);
+      return;
+    }
+    this.reconnectTimer = setTimeout(() => {
+      this.reconnectTimer = null;
+      this.connectWs();
+    }, this.backoffMs);
+    this.backoffMs = Math.min(this.backoffMs * 1.5, 10000);
+  }
+
+  private emitConnectionStateChange(): void {
+    const state = this.getConnectionState();
+    for (const listener of this.connectionStateListeners) {
+      try {
+        listener(state);
+      } catch {
+        // ignore listener errors
+      }
+    }
+  }
+
+  /** Get the current WebSocket connection state. */
+  getConnectionState(): ConnectionStateInfo {
+    return {
+      state: this.connectionState,
+      reconnectAttempt: this.reconnectAttempt,
+      maxReconnectAttempts: this.maxReconnectAttempts,
+      disconnectedAt: this.disconnectedAt,
+    };
+  }
+
+  /** Subscribe to connection state changes. Returns an unsubscribe function. */
+  onConnectionStateChange(
+    listener: (state: ConnectionStateInfo) => void,
+  ): () => void {
+    this.connectionStateListeners.add(listener);
+    return () => {
+      this.connectionStateListeners.delete(listener);
+    };
+  }
+
+  /** Reset connection state and restart reconnection attempts. */
+  resetConnection(): void {
+    this.reconnectAttempt = 0;
+    this.disconnectedAt = null;
+    this.connectionState = "disconnected";
+    if (this.reconnectTimer) {
+      clearTimeout(this.reconnectTimer);
+      this.reconnectTimer = null;
+    }
+    this.backoffMs = 500;
+    this.emitConnectionStateChange();
+    this.connectWs();
+  }
+
+  /** Send an arbitrary JSON message over the WebSocket connection. */
+  sendWsMessage(data: Record<string, unknown>): void {
+    const payload = JSON.stringify(data);
+    if (this.ws?.readyState === WebSocket.OPEN) {
+      this.ws.send(payload);
+      return;
+    }
+
+    // Keep only the newest active-conversation update while disconnected.
+    if (data.type === "active-conversation") {
+      this.wsSendQueue = this.wsSendQueue.filter((queued) => {
+        try {
+          const parsed = JSON.parse(queued) as { type?: unknown };
+          return parsed.type !== "active-conversation";
+        } catch {
+          return true;
+        }
+      });
+    }
+
+    if (this.wsSendQueue.length >= this.wsSendQueueLimit) {
+      const droppedType = typeof data.type === "string" ? data.type : "unknown";
+      console.warn("[ws] send queue full - dropping:", droppedType);
+      this.wsSendQueue.shift();
+    }
+    this.wsSendQueue.push(payload);
+
+    if (!this.ws || this.ws.readyState === WebSocket.CLOSED) {
+      this.connectWs();
+    }
+  }
+
+  onWsEvent(type: string, handler: WsEventHandler): () => void {
+    if (!this.wsHandlers.has(type)) {
+      this.wsHandlers.set(type, new Set());
+    }
+    this.wsHandlers.get(type)?.add(handler);
+    return () => {
+      this.wsHandlers.get(type)?.delete(handler);
+    };
+  }
+
+  disconnectWs(): void {
+    if (this.reconnectTimer) {
+      clearTimeout(this.reconnectTimer);
+      this.reconnectTimer = null;
+    }
+    this.ws?.close();
+    this.ws = null;
+    this.wsSendQueue = [];
+    // Reset connection state on intentional disconnect
+    this.reconnectAttempt = 0;
+    this.disconnectedAt = null;
+    this.connectionState = "disconnected";
+    this.emitConnectionStateChange();
+  }
+
+  // --- Text normalization helpers (used by chat domain methods) ---
+
+  protected normalizeAssistantText(text: string): string {
+    const stripped = stripAssistantStageDirections(text);
+    const trimmed = stripped.trim();
+    if (trimmed.length === 0) {
+      if (
+        text.trim().length === 0 ||
+        /^\(?no response\)?$/i.test(text.trim())
+      ) {
+        return GENERIC_NO_RESPONSE_TEXT;
+      }
+      return "";
+    }
+    if (/^\(?no response\)?$/i.test(trimmed)) {
+      return GENERIC_NO_RESPONSE_TEXT;
+    }
+    return trimmed;
+  }
+
+  protected normalizeGreetingText(text: string): string {
+    const stripped = stripAssistantStageDirections(text);
+    const trimmed = stripped.trim();
+    if (trimmed.length === 0 || /^\(?no response\)?$/i.test(trimmed)) {
+      return "";
+    }
+    return trimmed;
+  }
+
+  // --- Streaming chat endpoint (used by chat domain methods) ---
+
+  protected async streamChatEndpoint(
+    path: string,
+    text: string,
+    onToken: (token: string, accumulatedText?: string) => void,
+    channelType: ConversationChannelType = "DM",
+    signal?: AbortSignal,
+    images?: ImageAttachment[],
+    conversationMode?: ConversationMode,
+  ): Promise<{
+    text: string;
+    agentName: string;
+    completed: boolean;
+    usage?: ChatTokenUsage;
+  }> {
+    const res = await this.rawRequest(path, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        Accept: "text/event-stream",
+      },
+      body: JSON.stringify({
+        text,
+        channelType,
+        ...(images?.length ? { images } : {}),
+        ...(conversationMode ? { conversationMode } : {}),
+      }),
+      signal,
+    });
+
+    if (!res.body) {
+      throw new Error("Streaming not supported by this browser");
+    }
+
+    const decoder = new TextDecoder();
+    const reader = res.body.getReader();
+    let buffer = "";
+    let fullText = "";
+    let doneText: string | null = null;
+    let doneAgentName: string | null = null;
+    let doneUsage: ChatTokenUsage | undefined;
+    let receivedDone = false;
+
+    const findSseEventBreak = (
+      chunkBuffer: string,
+    ): { index: number; length: number } | null => {
+      const lfBreak = chunkBuffer.indexOf("\n\n");
+      const crlfBreak = chunkBuffer.indexOf("\r\n\r\n");
+
+      if (lfBreak === -1 && crlfBreak === -1) return null;
+      if (lfBreak === -1) return { index: crlfBreak, length: 4 };
+      if (crlfBreak === -1) return { index: lfBreak, length: 2 };
+      return lfBreak < crlfBreak
+        ? { index: lfBreak, length: 2 }
+        : { index: crlfBreak, length: 4 };
+    };
+
+    const parseDataLine = (line: string): void => {
+      const payload = line.startsWith("data:") ? line.slice(5).trim() : "";
+      if (!payload) return;
+
+      let parsed: {
+        type?: string;
+        text?: string;
+        fullText?: string;
+        agentName?: string;
+        message?: string;
+        usage?: {
+          promptTokens?: number;
+          completionTokens?: number;
+          totalTokens?: number;
+          model?: string;
+        };
+      };
+      try {
+        parsed = JSON.parse(payload) as typeof parsed;
+      } catch {
+        return;
+      }
+
+      if (!parsed.type && typeof parsed.text === "string") {
+        parsed.type = "token";
+      }
+
+      if (parsed.type === "token") {
+        const chunk = parsed.text ?? "";
+        const nextFullText =
+          typeof parsed.fullText === "string"
+            ? parsed.fullText
+            : chunk
+              ? mergeStreamingText(fullText, chunk)
+              : fullText;
+        if (nextFullText === fullText) return;
+        fullText = nextFullText;
+        onToken(chunk, fullText);
+        return;
+      }
+
+      if (parsed.type === "done") {
+        receivedDone = true;
+        if (typeof parsed.fullText === "string") doneText = parsed.fullText;
+        if (typeof parsed.agentName === "string" && parsed.agentName.trim()) {
+          doneAgentName = parsed.agentName;
+        }
+        if (parsed.usage) {
+          doneUsage = {
+            promptTokens: parsed.usage.promptTokens ?? 0,
+            completionTokens: parsed.usage.completionTokens ?? 0,
+            totalTokens: parsed.usage.totalTokens ?? 0,
+            model: parsed.usage.model,
+          };
+        }
+        // Terminal event: stop reading immediately instead of waiting for the
+        // server to close the body (some stacks leave the stream open briefly).
+        void reader.cancel("milady-sse-terminal-done").catch(() => {});
+        return;
+      }
+
+      if (parsed.type === "error") {
+        throw new Error(parsed.message ?? "generation failed");
+      }
+    };
+
+    // Contract: the API must emit `data: {"type":"done",...}` or
+    // `data: {"type":"error",...}` and then end the response. If the server
+    // stalls mid-stream (e.g. LLM provider timeout without error propagation),
+    // the idle timeout below aborts the read so the UI doesn't hang forever.
+    const SSE_IDLE_TIMEOUT_MS = 60_000;
+    while (true) {
+      let done = false;
+      let value: Uint8Array | undefined;
+      try {
+        const readPromise = reader.read();
+        const timeoutPromise = new Promise<never>((_, reject) => {
+          const id = setTimeout(
+            () => reject(new Error("SSE idle timeout — no data for 60s")),
+            SSE_IDLE_TIMEOUT_MS,
+          );
+          // Clear timeout if the read resolves first
+          void readPromise.finally(() => clearTimeout(id));
+        });
+        ({ done, value } = await Promise.race([readPromise, timeoutPromise]));
+      } catch (streamErr) {
+        console.warn("[api-client] SSE stream interrupted:", streamErr);
+        void reader.cancel("milady-sse-idle-timeout").catch(() => {});
+        break;
+      }
+      if (done || !value) break;
+
+      buffer += decoder.decode(value, { stream: true });
+      let eventBreak = findSseEventBreak(buffer);
+      while (eventBreak) {
+        const rawEvent = buffer.slice(0, eventBreak.index);
+        buffer = buffer.slice(eventBreak.index + eventBreak.length);
+        for (const line of rawEvent.split(/\r?\n/)) {
+          if (!line.startsWith("data:")) continue;
+          parseDataLine(line);
+        }
+        eventBreak = findSseEventBreak(buffer);
+      }
+    }
+
+    if (buffer.trim()) {
+      for (const line of buffer.split(/\r?\n/)) {
+        if (line.startsWith("data:")) parseDataLine(line);
+      }
+    }
+
+    const resolvedText = this.normalizeAssistantText(doneText ?? fullText);
+    return {
+      text: resolvedText,
+      agentName: doneAgentName ?? "Milady",
+      completed: receivedDone,
+      ...(doneUsage ? { usage: doneUsage } : {}),
+    };
+  }
+}
